requirements_railway.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
schedule==1.2.0

